organized tests into subdirectories, set up mocha.opts
