+++ src/app/core/device.js
@@ -1,6 +1,6 @@
 var values = require('./values.js');
-var uuid = require('node-uuid');
 var Params = values.Params;
+var Feature = require('./feature')
 
 /* The Device stores information about the physical properties of a given design. */
 class Device {
@@ -205,38 +205,6 @@ class Group {
 	}
 
 	//TODO: fromJSON()
-}
-
-class Feature {
-	constructor(type, params, name = new values.StringValue("New Feature"), id = Feature.generateID(), group = null){
-		if (id == undefined || name == undefined || type == undefined){
-			throw new Error("Cannot create feature with undefined values. id: "+  id + "name: "+ name + "type: " + type);
-		}
-		this.id = id;
-		this.name = name;
-		this.connections = {};
-		this.type = type;
-		this.params = params;
-		this.group = group;
-	}
-
-	static generateID(){
-		return uuid.v1();
-	}
-
-	toJSON(){
-		let output = {};
-		output.id = this.id;
-		output.name = this.name.toJSON();
-		output.type = this.type;
-		output.params = Params.toJSON(this.params);
-		//output.group = this.group.toJSON();
-		return output;
-	}
-
-	static fromJSON(json){
-		return new Feature(json.type, Params.fromJSON(json.params), json.name, json.id);
-	}
 }
 
 exports.Device = Device;
