+++ src/test/mocha.opts
@@ -1,0 +1,2 @@
+--recursive
+--compilers js:mocha-traceur
